Fix several warnings around lossy implicit conversion.

+++ src/cgame/default/views/MapListCollectionView.c
@@ -176,7 +176,7 @@ static void enumerateMaps(const char *path, void *data) {
 
 			GList *mapshots = cgi.Mapshots(path);
 
-			const size_t len = g_list_length(mapshots);
+			const guint len = g_list_length(mapshots);
 			if (len) {
 				const char *mapshot = g_list_nth_data(mapshots, Randomr(0, len));
